+++ .dev/build_latest.sh
@@ -1,2 +1,2 @@
-docker build -t pawelmalak/flame -t "pawelmalak/flame:$1" -f .docker/Dockerfile "$2" \
+docker build -t pawelmalak/flame -t "pawelmalak/flame:$1" -f .docker/Dockerfile . \
   && docker push pawelmalak/flame && docker push "pawelmalak/flame:$1"